Move override_csp_to_allow_acuant spec to acuant_concern_spec.rb
